split out queries for cart page
